applySwipeGesture is now in extensions package. Use viewType extension also for emptyItem.
