+++ recycleradapter/src/main/java/net/gotev/recycleradapter/Extensions.kt
@@ -1,0 +1,7 @@
+package net.gotev.recycleradapter
+
+/**
+ * @author Aleksandar Gotev
+ */
+internal fun AdapterItem<*>?.viewType() = this?.javaClass?.name?.hashCode() ?: 0
+internal fun Class<out AdapterItem<*>>.viewType() = hashCode()
